Removed deprecated combiner. Fixed circular dependency for packer & seed packages. Updated fixtures
